refactor: 페이지네이션 훅 삭제 및 코드통합 (FE-142)
